Linter + project structure
