--- src/TasksApi.js
@@ -1,10 +1,0 @@
-export default class TaskApi{
-    static async getTasks() {
-        return JSON.parse(localStorage.getItem("tasks") || "[]") ;
-    }
-
-    static async saveTasks(tasks) {
-        console.log(tasks)
-        localStorage.setItem("tasks", JSON.stringify(tasks));
-    }
-}
