+++ src/components/Header.svelte
@@ -1,7 +1,6 @@
 <script>
-  import {tasks} from '../tasks';
+  import { tasks } from '$lib/tasks';
   $: tasksCount = $tasks.length;
- 
 </script>
 
 <style>
@@ -13,11 +12,12 @@
     color: #eeeeee;
     font-weight: bold;
     font-size: 1.5em;
-    }
-    h3{
-      font-size: 0.9em;
-      opacity: 0.8;
-    }
+  }
+
+  h3 {
+    font-size: 0.9em;
+    opacity: 0.8;
+  }
 </style>
 
 <div class="header">
@@ -25,6 +25,4 @@
     <h2>Bienvenido a TimeTracker </h2>
     <h3> Cantidad de Tareas: {tasksCount} </h3>
   </div>
-
-  <!-- <span> {completedItemCount}/{itemCount} </span> -->
 </div>
